Refactor. 두 날짜 사이의 일을 계산하여 숫자로 변환하는 함수 유틸로 옮김
